// src/generated/multi-airport-city-lookup.ts
export interface components {
    schemas: {
        MultiAirportCityLookupResponse: {
            Cities?: components["schemas"]["MultiAirportCity"][];
            /** @description Array of links for different types of relations, e.g. self, linkTemplate */
            Links?: components["schemas"]["ResponseLink"][];
        };
        MultiAirportCity: {
            /**
             * @description Multi-airport city (MAC) code.
             * @example BER
             */
            code?: string;
            /**
             * @description City name that is associated with the city code in the `code` object.
             * @example Berlin
             */
            name?: string;
            /**
             * @description Country code that is associated with the city code in the `code` object.
             * @example DE
             */
            countryCode?: string;
            /**
             * @description The name of the country that corresponds to `countryCode`.
             * @example Germany
             */
            countryName?: string;
            /**
             * @description The name of the region that is assigned to the country.
             * @example Europe
             */
            regionName?: string;
            /** @description Array of links for different types of relations, e.g. airportsInCity */
            Links?: components["schemas"]["CityLink"][];
        };
        CityLink: {
            /**
             * @description Describes relationship between href and current request.
             * @example airportsInCity
             * @enum {string}
             */
            rel?: "airportsInCity";
            /**
             * @description Link to related API request.
             * @example https://api.cert.platform.sabre.com/v1/lists/supported/cities/BER/airports
             */
            href?: string;
        };
        ResponseLink: {
            /**
             * @description Describes relationship between href and current request.
             * @example self
             * @enum {string}
             */
            rel?: "self" | "linkTemplate";
            /**
             * @description Link to related API request.
             * @example https://api.cert.platform.sabre.com/v1/lists/supported/cities?country=DE
             */
            href?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}